fix : 2/2 프로그래머스 레벨 2: [프린터 문제]
주석추가
